# doc/openapi.json
{
  "components": {
    "schemas": {
...
      "RefreshTokenInput": {
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
...
      "TokenPair": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Make sure to include the X-ORG-ID header when using this API key.",
        "in": "header",
        "name": "X-API-KEY",
        "type": "apiKey"
      },
      "JWTAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@dot.industries",
      "name": "Support",
      "url": "https://app.brease.run/support"
    },
    "description": "Business rule engine as a service",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "brease API",
    "version": "0.1.0"
  },
  "openapi": "3.0.1",
  "paths": {
    "/refreshToken": {
      "post": {
        "description": "Refresh the short lived access token for web access",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Cloud hosted production server",
      "url": "https://api.brease.run"
    },
    {
      "description": "Development server",
      "url": "http://localhost:4400"
    }
  ],
...
}